src/models/content_based.py
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import linear_kernel

class ContentBasedRecommender:
    def __init__(self):
        self.vectorizer = TfidfVectorizer(stop_words='english')
        self.tfidf_matrix = None

    def fit(self, movies_df):
        """
        Fits the TF-IDF vectorizer on movie overviews.
        """
        self.tfidf_matrix = self.vectorizer.fit_transform(movies_df['overview'].fillna(''))

    def get_recommendations(self, movie_title, movies_df, top_k=10):
        # Placeholder logic
        return []
