requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
pydantic
pydantic-settings
python-dotenv
httpx
pytest
jinja2
python-multipart
motor
pymongo
pickle5 ; python_version < "3.8"
